+++ OctopusParent/Common/src/main/java/com/octopus/common/dao/domain/ProductBaseInfoDto.java
@@ -1,6 +1,7 @@
 package com.octopus.common.dao.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 文件创建时写入注释内容
@@ -13,15 +14,15 @@ public class ProductBaseInfoDto implements Serializable {
     String productId;
     String productRiskLevel;
     String productType;
-    float productRaiseAmount;
-    float productRemainAmount;
+    BigDecimal productRaiseAmount;
+    BigDecimal productRemainAmount;
     String productName;
     String registarCode;
     String registarName;
 
 
-    public ProductBaseInfoDto(String productId, String productRiskLevel, String productType, float productRaiseAmount,
-                              float productRemainAmount, String productName, String registarCode, String registarName) {
+    public ProductBaseInfoDto(String productId, String productRiskLevel, String productType, BigDecimal productRaiseAmount,
+                              BigDecimal productRemainAmount, String productName, String registarCode, String registarName) {
         this.productId = productId;
         this.productRiskLevel = productRiskLevel;
         this.productType = productType;
@@ -56,19 +57,19 @@ public class ProductBaseInfoDto implements Serializable {
         this.productType = productType;
     }
 
-    public float getProductRaiseAmount() {
+    public BigDecimal getProductRaiseAmount() {
         return productRaiseAmount;
     }
 
-    public void setCertificationType(float productRaiseAmount) {
+    public void setCertificationType(BigDecimal productRaiseAmount) {
         this.productRaiseAmount = productRaiseAmount;
     }
 
-    public float getProductRemainAmount() {
+    public BigDecimal getProductRemainAmount() {
         return productRemainAmount;
     }
 
-    public void setProductRemainAmount(float productRemainAmount) {
+    public void setProductRemainAmount(BigDecimal productRemainAmount) {
         this.productRemainAmount = productRemainAmount;
     }
