+++ OctopusParent/Common/src/main/java/com/octopus/common/dao/mapper/PaymentInfoMapper.java
@@ -49,7 +49,7 @@ public interface PaymentInfoMapper {
             "WHERE order_seq = #{orderSeq} " +
             "  AND pay_type = #{payType} " +
             " ")
-    Long delete(@Param("orderSeq")float  orderSeq,@Param("payType")String payType);
+    Long delete(@Param("orderSeq")double  orderSeq,@Param("payType")String payType);
 
     /**
      * 查询所有
@@ -66,5 +66,5 @@ public interface PaymentInfoMapper {
      * @return
      */
     @Select("select * from t_payment_info where  order_seq = #{orderSeq}  AND pay_type = #{payType}" )
-    PaymentInfoDto selectById(@Param("orderSeq")float  orderSeq,@Param("payType")String payType);
+    PaymentInfoDto selectById(@Param("orderSeq")double  orderSeq,@Param("payType")String payType);
 }
